Mitigada vulnerabilidad de Broken Access Control para tours e imágenes.
- Se ha creado un método isAdmin() en UserService
- Se ha utilizado este método en los controladores Rest de Image y de Tour para todos los métodos put y delete.

+++ src/main/java/es/codeurjc/daw/library/controller/rest/ImageRestController.java
@@ -3,11 +3,16 @@ package es.codeurjc.daw.library.controller.rest;
 import es.codeurjc.daw.library.dto.ImageDTO;
 import es.codeurjc.daw.library.dto.ImageMapper;
 import es.codeurjc.daw.library.model.Image;
+import es.codeurjc.daw.library.model.User;
 import es.codeurjc.daw.library.service.ImageService;
+import es.codeurjc.daw.library.service.UserService;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.io.IOException;
 import java.net.URI;
@@ -23,6 +28,9 @@ public class ImageRestController {
 
     @Autowired
     private ImageMapper imageMapper;
+
+    @Autowired
+    private UserService userService;
 
     @GetMapping("/{id}")
     public ResponseEntity<ImageDTO> getImage(@PathVariable long id) {
@@ -43,7 +51,7 @@ public class ImageRestController {
 
         return ResponseEntity
                 .ok()
-                .header("Content-Type", "image/jpeg") 
+                .header("Content-Type", "image/jpeg")
                 .body(imageBytes);
     }
 
@@ -70,6 +78,13 @@ public class ImageRestController {
             @PathVariable long id,
             @RequestParam MultipartFile imageFile) throws IOException {
 
+        // Broken Access Control
+        User user = userService.getLoggedUser();
+
+        if (!userService.isAdmin(user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         if (imageFile.isEmpty()) {
             return ResponseEntity.badRequest().build();
         }
@@ -80,6 +95,14 @@ public class ImageRestController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<ImageDTO> deleteImage(@PathVariable long id) {
+
+        // Broken Access Control
+        User user = userService.getLoggedUser();
+
+        if (!userService.isAdmin(user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+
         Image image = imageService.deleteImage(id);
         return ResponseEntity.ok(imageMapper.toDTO(image));
     }
